requirements.txt
a2a==0.44
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.4
aiohttp-jinja2==1.6
aiosignal==1.3.2
alembic==1.14.1
alpha_vantage==3.0.0
annotated-types==0.7.0
anthropic==0.56.0
anyio==4.9.0
arrow==1.3.0
attrs==25.3.0
audioop-lts==0.2.1
Automat==25.4.16
bcrypt==4.3.0
beautifulsoup4==4.13.4
biiopython==0.0.5
black==25.1.0
blinker==1.9.0
boto3==1.39.1
botocore==1.39.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
coverage==7.9.2
cryptography==43.0.1
cssselect==1.3.0
curl_cffi==0.11.4
decorator==5.2.1
deep-translator==1.11.4
defusedxml==0.7.1
distro==1.9.0
fastapi==0.115.14
feedparser==6.0.11
filelock==3.18.0
flake8==7.3.0
Flask==2.0.2
flask-cors==6.0.0
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.5.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.0.3
jinja2-time==0.2.0
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.67
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.4.4
loguru==0.7.2
lxml==6.0.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
multitasking==0.0.11
mypy==1.16.1
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.1
numpy==2.2.6
openai==1.93.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parsel==1.10.0
pathspec==0.12.1
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.2
Protego==0.5.0
protobuf==3.20.3
public==2020.12.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.9.0
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyDispatcher==2.0.7
pyflakes==3.4.0
Pygments==2.19.2
pyjabber==0.3.0
pyOpenSSL==25.1.0
pytest==8.4.1
pytest-cov==6.2.1
python-a2a==0.5.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2022.1
PyYAML==6.0.2
queuelib==1.8.0
redis==6.2.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
responses==0.25.7
rich==14.0.0
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
Scrapy==2.13.3
self==2020.12.3
service-identity==24.2.0
setuptools==80.9.0
sgmllib3k==1.0.0
singletonify==0.2.4
six==1.17.0
slixmpp==1.9.1
sniffio==1.3.1
soupsieve==2.7
spade==4.1.0
SpeechRecognition==3.14.3
SQLAlchemy==2.0.40
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
textblob==0.19.0
threadpoolctl==3.6.0
timeago==1.0.16
tldextract==5.3.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
Twisted==25.5.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.35.0
w3lib==2.3.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.3.7
win32_setctime==1.2.0
winloop==0.1.8
yarl==1.20.1
yfinance==0.2.64
zope.interface==7.2
zstandard==0.23.0
